requirements.txt
selenium
psycopg2-binary
webdriver_manager
lxml